refactor: 컴포넌트명 GNB를 NavBar로 변경
