+++ src/components/NavBar.styles.ts
