+++ src/components/NavBar.tsx
@@ -1,9 +1,9 @@
-import { useEffect, useRef, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import SideMenu from './SideMenu';
-import * as S from "./GNB.styles";
+import { useEffect, useRef, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import SideMenu from "./SideMenu";
+import * as S from "./NavBar.styles";
 
-const GNB = () => {
+const NavBar = () => {
   const navigate = useNavigate();
   const [isMenuOpen, setMenuOpen] = useState(false);
 
@@ -15,15 +15,17 @@ const GNB = () => {
   };
 
   const handleLogoClick = () => {
-    setMenuOpen(false)
-    navigate('/');
-  }
+    setMenuOpen(false);
+    navigate("/");
+  };
 
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
       if (
-        sideMenuRef.current && !sideMenuRef.current.contains(event.target as Node) &&
-        parentRef.current && !parentRef.current.contains(event.target as Node)
+        sideMenuRef.current &&
+        !sideMenuRef.current.contains(event.target as Node) &&
+        parentRef.current &&
+        !parentRef.current.contains(event.target as Node)
       ) {
         setMenuOpen(false);
       }
@@ -47,4 +49,4 @@ const GNB = () => {
   );
 };
 
-export default GNB;
+export default NavBar;
